More speed and size optimizations for multiplexed mode
Calculate toggleRows[row] before setting BLANK high, and store it in a register,
that way BLANK is high for a shorter amount of time. Also, move the increment of
row after setting BLANK low, that way BLANK is high for an even shorter amount
of time.

Now, the important logic that must run every 4096 clocks is completely
deterministic.