Only 5 wires (rather than 7) are needed for full dot correction support when multiplexing
Also made sure that PB2 is configured as an output if TLC5940_SPI_MODE = 0.

Added TLC5940_DCPRG_HARDWIRED_TO_VCC configuration option, that if set means you don't
have to waste a pin for DCPRG, as long as the TLC5940's DCPRG pin is hardwired to VCC,
and you always manually load in the values. This is also useful for driving a TLC5944,
since it doesn't have a DCPRG pin.

Added TLC5940_XLAT_AND_BLANK_HARDWIRED_TOGETHER configuration option, that if set means
you don't have to waste a pin for BLANK, as long as you are multiplexing, and have
connected the TLC5940's XLAT and BLANK pins together (with the pullup resistor on
BLANK, now on both BLANK and XLAT).